// src/typings/Anime.ts
export enum AnimePgaRatings {
    None = 'none',
    G = 'g',
    PG = 'pg',
    PG_13 = 'pg_13',
    R = 'r',
    R_plus = 'r_plus',
    RX = 'rx',
}

export enum Seasons {
    winter = 'winter',
    spring = 'spring',
    summer = 'summer',
    autumn = 'autumn',
}

export enum AnimeMediaTypes {
    tv = 'tv',
    tv_special = 'tv_special',
    special = 'special',
    ona = 'ona',
    ova = 'ova',
    movie = 'movie',
}

export enum NotifyStatuses {
    AnimeRelease = 'anime_released',
    EpisodeRelease = 'episode_released',
    FinalEpisodeReleased = 'final_episode_released',
}

export enum AnimeListStatuses {
    planned = 'planned',
    watching = 'watching',
    rewatching = 'rewatching',
    completed = 'completed',
    on_hold = 'on_hold',
    dropped = 'dropped',
}
// ...
export interface ShortAnime {
    id: number;
    image: string;
    name: string;
    description: string;
    status: string;
    slug: string;
    currentEpisodes: number;
    englishName: string;
    firstEpisodeAired: string;
    franchiseName: string;
    japaneseName: string;
    kodikLink: string;
    lastEpisodeAired: string;
    maxEpisodes: number;
    mediaType: AnimeMediaTypes;
    rpaRating: AnimePgaRatings;
    shikimoriId: number;
    shikimoriRating: number;
    banned: boolean;
    censored: boolean;
    hasRelation: boolean;
    rating: number;
    season: string;
}
// ...
export interface WatchList {
    id: number;
    animeId: number;
    userId: number;
    status: AnimeListStatuses;
    isFavorite: boolean;
    watchedEpisodes: number;
    rating: number;
    shikimoriId?: number;
}

export interface WatchListWithAnime extends WatchList {
    anime: ShortAnime;
}

enum NotificationStatuses {
    EpisodeRelease = 'episode_released',
    FinalEpisodeReleased = 'final_episode_released',
    AnimeReleased = 'anime_released'

}

export interface UserNotification {
    id: number,
    animeId: number,
    status: NotificationStatuses,
    userId: number,
    episode: number,
    groupId: number,
    isRead: boolean,
    createdAt: string
}

export interface AnimeAnnouncement {
    animeName: string,
    groupType?: GroupType, 
    groupName?: string,
    episode: number,
    maxEpisodes: number,
    notificationType: NotifyStatuses,
    slug: string,
    image: string,
    userIds: number[]
}


export enum GroupType{
    Voice = "voice",
    Subtitles = "subtitles"
}